central-server/backend_v5/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
mysql-connector-python==8.2.0
APScheduler==3.10.4
python-dotenv==1.0.0
gunicorn==21.2.0
